// PLANTILLAS/admindashboard/src/app/Models/Clientes.ts
export class Clientes{
    clie_ID: any;
    clie_Nombres: any;
    clie_Apellidos: any;
    clie_DNI: any;
    clie_Sexo: any;
    clie_Telefono: any;
    clie_Habilitado: any;
    clie_Estado: any;
    clie_UsuarioCreador: any;
    usu_Creador: any;
    clie_FechaCreacion: any;
    clie_UsuarioModificador: any;
    usu_Modificador: any;
    clie_FechaModificacion: any;
    empl_crea: any;
    empl_Modifica: any;
}